# Add FingerTree-like push_right, drop_left for performance; move out Empty variant