--- a/benches/creating.rs
+++ b/benches/creating.rs
@@ -13,6 +13,16 @@ fn criterion_benchmark(c: &mut Criterion) {
     })
   });
 
+  c.bench_function("creating list disable balancing", |b| {
+    b.iter(|| {
+      let mut data = TernaryTreeList::Empty;
+
+      for idx in 0..1000 {
+        data = data.append(idx, true)
+      }
+    })
+  });
+
   c.bench_function("push_right list", |b| {
     b.iter(|| {
       let mut data = TernaryTreeList::Empty;
@@ -27,20 +37,10 @@ fn criterion_benchmark(c: &mut Criterion) {
     b.iter(|| {
       let mut data = TernaryTreeList::Empty;
 
+      // TODO overflowed
       for idx in 0..1000 {
         let pos = idx / 2;
         data = data.insert(pos, idx, false).unwrap()
-      }
-    })
-  });
-
-  c.bench_function("inserting in middle without balancing", |b| {
-    b.iter(|| {
-      let mut data = TernaryTreeList::Empty;
-
-      for idx in 0..1000 {
-        let pos = idx / 2;
-        data = data.insert(pos, idx, true).unwrap()
       }
     })
   });
@@ -49,8 +49,7 @@ fn criterion_benchmark(c: &mut Criterion) {
     let mut data = TernaryTreeList::Empty;
 
     for idx in 0..1000 {
-      let pos = idx / 2;
-      data = data.insert(pos, idx, false).unwrap()
+      data = data.push(idx);
     }
 
     b.iter(move || {
@@ -62,12 +61,59 @@ fn criterion_benchmark(c: &mut Criterion) {
     })
   });
 
+  c.bench_function("rest from push_right", |b| {
+    let mut data = TernaryTreeList::Empty;
+
+    for idx in 0..1000 {
+      data = data.push_right(idx);
+    }
+
+    b.iter(move || {
+      let mut d = data.to_owned();
+
+      while !d.is_empty() {
+        d = d.slice(1, d.len()).unwrap()
+      }
+    })
+  });
+
+  c.bench_function("drop-left", |b| {
+    let mut data = TernaryTreeList::Empty;
+
+    for idx in 0..1000 {
+      data = data.push(idx);
+    }
+
+    b.iter(move || {
+      let mut d = data.to_owned();
+
+      while d.len() > 1 {
+        d = d.drop_left()
+      }
+    })
+  });
+
+  c.bench_function("drop-left from push_right", |b| {
+    let mut data = TernaryTreeList::Empty;
+
+    for idx in 0..1000 {
+      data = data.push_right(idx);
+    }
+
+    b.iter(move || {
+      let mut d = data.to_owned();
+
+      while d.len() > 1 {
+        d = d.drop_left()
+      }
+    })
+  });
+
   c.bench_function("slice", |b| {
     let mut data = TernaryTreeList::Empty;
 
     for idx in 0..1000 {
-      let pos = idx / 2;
-      data = data.insert(pos, idx, false).unwrap()
+      data = data.push(idx)
     }
 
     b.iter(move || {