--- a/examples/push.rs
+++ b/examples/push.rs
@@ -1,0 +1,7 @@
+use im_ternary_tree::TernaryTreeList;
+
+pub fn main() -> Result<(), String> {
+  let tree = TernaryTreeList::new();
+
+  Ok(())
+}