--- a/examples/push.rs
+++ b/examples/push.rs
@@ -1,7 +1,12 @@
 use im_ternary_tree::TernaryTreeList;
 
 pub fn main() -> Result<(), String> {
-  let tree = TernaryTreeList::new();
+  let mut tree: TernaryTreeList<usize> = TernaryTreeList::Empty;
+
+  for idx in 0..60 {
+    tree = tree.push_right(idx);
+    println!("{}", tree.format_inline());
+  }
 
   Ok(())
 }