--- a/src/slice.rs
+++ b/src/slice.rs
@@ -2,20 +2,26 @@ use std::fmt::{Debug, Display};
 use std::hash::Hash;
 use std::sync::Arc;
 
+use crate::tree::*;
 use crate::TernaryTreeList;
-use crate::TernaryTreeList::*;
+
+use crate::tree::TernaryTree::*;
 
 impl<T> From<Vec<T>> for TernaryTreeList<T>
 where
   T: Clone + Display + Eq + PartialEq + Debug + Ord + PartialOrd + Hash,
 {
   fn from(xs: Vec<T>) -> Self {
-    let mut ys: Vec<Self> = Vec::with_capacity(xs.len());
-    for x in &xs {
-      ys.push(Leaf(Arc::new(x.to_owned())))
-    }
+    if xs.is_empty() {
+      TernaryTreeList::Empty
+    } else {
+      let mut ys: Vec<TernaryTree<T>> = Vec::with_capacity(xs.len());
+      for x in &xs {
+        ys.push(Leaf(Arc::new(x.to_owned())))
+      }
 
-    Self::rebuild_list(xs.len(), 0, &ys)
+      TernaryTreeList::Tree(TernaryTree::rebuild_list(xs.len(), 0, &ys))
+    }
   }
 }
 
@@ -24,12 +30,16 @@ where
   T: Clone + Display + Eq + PartialEq + Debug + Ord + PartialOrd + Hash,
 {
   fn from(xs: &Vec<T>) -> Self {
-    let mut ys: Vec<Self> = Vec::with_capacity(xs.len());
-    for x in xs {
-      ys.push(Leaf(Arc::new((*x).to_owned())))
-    }
+    if xs.is_empty() {
+      TernaryTreeList::Empty
+    } else {
+      let mut ys: Vec<TernaryTree<T>> = Vec::with_capacity(xs.len());
+      for x in xs {
+        ys.push(Leaf(Arc::new(x.to_owned())))
+      }
 
-    Self::rebuild_list(xs.len(), 0, &ys)
+      TernaryTreeList::Tree(TernaryTree::rebuild_list(xs.len(), 0, &ys))
+    }
   }
 }
 
@@ -39,10 +49,15 @@ where
   T: Clone + Display + Eq + PartialEq + Debug + Ord + PartialOrd + Hash,
 {
   fn from(xs: &[T; N]) -> Self {
-    let mut ys: Vec<Self> = Vec::with_capacity(xs.len());
-    for x in xs {
-      ys.push(Leaf(Arc::new((*x).to_owned())))
+    if xs.is_empty() {
+      TernaryTreeList::Empty
+    } else {
+      let mut ys: Vec<TernaryTree<T>> = Vec::with_capacity(xs.len());
+      for x in xs {
+        ys.push(Leaf(Arc::new(x.to_owned())))
+      }
+
+      TernaryTreeList::Tree(TernaryTree::rebuild_list(xs.len(), 0, &ys))
     }
-    Self::rebuild_list(xs.len(), 0, &ys)
   }
 }