--- a/src/slice.rs
+++ b/src/slice.rs
@@ -1,6 +1,5 @@
 use std::fmt::{Debug, Display};
 use std::hash::Hash;
-use std::sync::Arc;
 
 use crate::tree::*;
 use crate::TernaryTreeList;
@@ -17,7 +16,7 @@ where
     } else {
       let mut ys: Vec<TernaryTree<T>> = Vec::with_capacity(xs.len());
       for x in &xs {
-        ys.push(Leaf(Arc::new(x.to_owned())))
+        ys.push(Leaf(x.to_owned()))
       }
 
       TernaryTreeList::Tree(TernaryTree::rebuild_list(xs.len(), 0, &ys))
@@ -35,7 +34,7 @@ where
     } else {
       let mut ys: Vec<TernaryTree<T>> = Vec::with_capacity(xs.len());
       for x in xs {
-        ys.push(Leaf(Arc::new(x.to_owned())))
+        ys.push(Leaf(x.to_owned()))
       }
 
       TernaryTreeList::Tree(TernaryTree::rebuild_list(xs.len(), 0, &ys))
@@ -54,7 +53,7 @@ where
     } else {
       let mut ys: Vec<TernaryTree<T>> = Vec::with_capacity(xs.len());
       for x in xs {
-        ys.push(Leaf(Arc::new(x.to_owned())))
+        ys.push(Leaf(x.to_owned()))
       }
 
       TernaryTreeList::Tree(TernaryTree::rebuild_list(xs.len(), 0, &ys))