Based on experiences from FingerTrees, shallow branches at head and tail does help in reducing the overhead of adding/dropping elements. So I'm trying to do something similar in this library to add `.push_right` `.drop_left`. Very roughly 50% time reduced in both operations. Of cause that performance of `.drop_left` highly depends on the original structure since ternary tree is highly chaotic in its layout.

Also `Empty` is now moved out and only allowed as a variant at top level, which also decreased overhead of branching lookup.

Push right:

```
0
(0 1)
(0 1 2)
((0 1 2) 3)
((0 1 2) (3 4))
((0 1 2) (3 4 5))
((0 1 2) (3 4 5) 6)
((0 1 2) (3 4 5) (6 7))
((0 1 2) (3 4 5) (6 7 8))
(((0 1 2) (3 4 5) (6 7 8)) 9)
(((0 1 2) (3 4 5) (6 7 8)) (9 10))
(((0 1 2) (3 4 5) (6 7 8)) (9 10 11))
(((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) 12))
(((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13)))
(((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14)))
(((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) 15))
(((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16)))
(((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)))
(((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18)
((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) 19)
((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) (19 20))
((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) (19 20 21))
((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) 22))
((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23)))
((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24)))
((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) 25))
((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26)))
((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)))
((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28)
(((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) 29)
(((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) (29 30))
(((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) (29 30 31))
(((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) 32))
(((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33)))
(((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34)))
(((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) 35))
(((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36)))
(((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)))
(((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38)
((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) 39)
((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) (39 40))
((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) (39 40 41))
((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) 42))
((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43)))
((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44)))
((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) 45))
((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46)))
((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)))
((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48)
(((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) 49)
(((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) (49 50))
(((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) (49 50 51))
(((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) ((49 50 51) 52))
(((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) ((49 50 51) (52 53)))
(((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) ((49 50 51) (52 53 54)))
(((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) ((49 50 51) (52 53 54) 55))
(((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) ((49 50 51) (52 53 54) (55 56)))
(((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) ((49 50 51) (52 53 54) (55 56 57)))
(((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) ((49 50 51) (52 53 54) (55 56 57)) 58)
((((((((0 1 2) (3 4 5) (6 7 8)) ((9 10 11) (12 13 14) (15 16 17)) 18) ((19 20 21) (22 23 24) (25 26 27)) 28) ((29 30 31) (32 33 34) (35 36 37)) 38) ((39 40 41) (42 43 44) (45 46 47)) 48) ((49 50 51) (52 53 54) (55 56 57)) 58) 59)
```

Drop left:

```
(((948 949 950) 951 ((952 953 954) 955 (956 957 958))) 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
(((949 950) 951 ((952 953 954) 955 (956 957 958))) 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
((950 951 ((952 953 954) 955 (956 957 958))) 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
((951 ((952 953 954) 955 (956 957 958))) 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
(((952 953 954) 955 (956 957 958)) 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
(((953 954) 955 (956 957 958)) 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
((954 955 (956 957 958)) 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
((955 (956 957 958)) 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
((956 957 958) 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
((957 958) 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
(958 959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
(959 (((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999)))
(((((960 961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((((961 962) 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
((((962 963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
((((963 (964 965 966)) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
((((964 965 966) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
((((965 966) 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((966 967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((967 ((968 969 970) 971 (972 973 974))) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
((((968 969 970) 971 (972 973 974)) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
((((969 970) 971 (972 973 974)) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((970 971 (972 973 974)) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((971 (972 973 974)) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((972 973 974) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((973 974) 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
((974 975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
((975 (((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990)))) 991 (((992 993 994) 995 (996 997 998)) 999))
((((976 977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990))) 991 (((992 993 994) 995 (996 997 998)) 999))
((((977 978) 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((978 979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((979 (980 981 982)) 983 ((984 985 986) 987 (988 989 990))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((980 981 982) 983 ((984 985 986) 987 (988 989 990))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((981 982) 983 ((984 985 986) 987 (988 989 990))) 991 (((992 993 994) 995 (996 997 998)) 999))
((982 983 ((984 985 986) 987 (988 989 990))) 991 (((992 993 994) 995 (996 997 998)) 999))
((983 ((984 985 986) 987 (988 989 990))) 991 (((992 993 994) 995 (996 997 998)) 999))
(((984 985 986) 987 (988 989 990)) 991 (((992 993 994) 995 (996 997 998)) 999))
(((985 986) 987 (988 989 990)) 991 (((992 993 994) 995 (996 997 998)) 999))
((986 987 (988 989 990)) 991 (((992 993 994) 995 (996 997 998)) 999))
((987 (988 989 990)) 991 (((992 993 994) 995 (996 997 998)) 999))
((988 989 990) 991 (((992 993 994) 995 (996 997 998)) 999))
((989 990) 991 (((992 993 994) 995 (996 997 998)) 999))
(990 991 (((992 993 994) 995 (996 997 998)) 999))
(991 (((992 993 994) 995 (996 997 998)) 999))
(((992 993 994) 995 (996 997 998)) 999)
((993 994) (995 (996 997 998)) 999)
(994 (995 (996 997 998)) 999)
(995 (996 997 998) 999)
(996 (997 998) 999)
(997 998 999)
(998 999)
999
```

A rough comparison to FingerTree:

```
    Finished bench [optimized] target(s) in 7.78s
     Running unittests (target/release/deps/ds_perf-8b76c1047a1696f5)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/perf-63ee6f925ce067ad)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
ternary tree pushing    time:   [3.8072 ms 3.8724 ms 3.9553 ms]
                        change: [-0.7091% +1.4926% +4.0533%] (p = 0.23 > 0.05)
                        No change in performance detected.
Found 15 outliers among 100 measurements (15.00%)
  2 (2.00%) high mild
  13 (13.00%) high severe

finger tree pushing     time:   [6.1121 ms 6.1613 ms 6.2238 ms]
                        change: [-0.0959% +1.1011% +2.3405%] (p = 0.07 > 0.05)
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) high mild
  8 (8.00%) high severe

ternary tree init       time:   [2.5574 ms 2.5656 ms 2.5759 ms]
                        change: [-7.5446% -4.7372% -2.3178%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe

finger tree init        time:   [1.9776 ms 1.9814 ms 1.9860 ms]
                        change: [+0.1007% +0.4458% +0.7760%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe

ternary tree rest       time:   [5.0341 ms 5.0476 ms 5.0644 ms]
                        change: [-1.1394% -0.5841% -0.0515%] (p = 0.04 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low mild
  7 (7.00%) high mild
  4 (4.00%) high severe

finger tree rest        time:   [3.8543 ms 3.8901 ms 3.9309 ms]
                        change: [+0.3233% +1.6221% +2.9508%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe

ternary tree access iter
                        time:   [450.65 us 453.37 us 456.99 us]
                        change: [-4.9965% -4.2309% -3.3939%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  4 (4.00%) high mild
  10 (10.00%) high severe

finger tree access iter time:   [45.597 ms 45.715 ms 45.871 ms]
                        change: [+0.1407% +0.4670% +0.8375%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
```